Change type of ids to long for Route model  (storage)

+++ app/src/main/java/casak/ru/geofencer/storage/model/Route.java
@@ -21,7 +21,7 @@ public class Route extends BaseModel {
 
     @Column
     @Unique
-    public int fieldId;
+    public long fieldId;
 
     @ForeignKey(tableClass = RouteType.class)
     @Column
